fix: split up auth tests
